refactor(allocator): move `Allocator` into own module (#8656)
Pure refactor. `oxc_allocator` crate is getting quite crowded now. Move `Allocator` into a separate file.
